fix: battery_soc type int→float (Sungrow returns 3.8)

+++ app/models.py
@@ -95,7 +95,7 @@ class ChatRequest(BaseModel):
 class EnergyInfo(BaseModel):
     power_w: int = 0
     daily_solar_kwh: float = 0.0
-    battery_soc: int = 0
+    battery_soc: float = 0.0
 
 
 class DinnerInfo(BaseModel):
